I have no idea what I did
Alright, so I'm mostly doing this commit because I feel like I haven't
done enough commits lately, but I also want to share something…strange.
I'm not sure what I did to cause this, but the window seems to be
generating semi-random rectangles.  There is, however, a certain amount
of consistency, as seen by the fact that the rectangle in the top-left
is always that particular shade of…cyan?  and the black rectangles seem
to be always in the same place.  I'll attach some images below.

Also, this commit adds the assets file, which needs to be put in
`~/.ExNihilo/assets/` in order for the program to work.  I also changed
PLANS.md again, although I don't know how.