Nothing as permanent as a temporary solution
I tried for far too long to get the system I had in mind to work.  I
couldn't figure it out at all.  For now, I have temporarily decided to
resort to BMPs, since I cannot work on just this one specific problem
for so long, as this is destroying my soul, for there are far more
interesting things I want to do.

In other news, I've completely reorganized the file structure:  With the
exception of the source code (`Ex Nihilo.c`) and the README, all files
are now stored either in `docs/` or in `Assets/`, depending on what type
of file they are.

 dfaij;nhgs; ohg
;;;;;;;;;;;;nnnnjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjjsnnnnnnnnn