Basic text logic implemented
Let me first state this:  Yes, that massive switch statement was
necessary.  No, the solution that you're thinking of did not work.  I
have no idea why either.  This is also terribly undercommented, a
problem which I do plan to fix eventually.  For now, however, I intend
to work on the assets for the text logic and expand PLANS.md with some
game mechanics that will make this unique.  This may take a bit, but it
shouldn't take nearly as long as the mess I just had to deal with.