Made time work and fullscreen optional
Apologies that this took so long:  I've been busy with a lot of other
things, which doesn't compound well with being a lazy prick.

The game loop will now only do things every nifth of a second, and
there's now a popup window to ask if you want fullscreen which
convieniently is the size of the window if you _don't_ want fullscreen.