Fixed text drawing
There was a bug where text was being shifted over to the right by one position.  I tried fixing this with negative numbers and caused a
segfault, so I implemented a one-time check that will occur everytime.

In addition, text can now be restricted to a specific area of the
screen, instead of just going towards the edge.  So that's nice.