Made half-a-dozen characters have transparency
I'm going to save the rest of them for later, partly because I'm a lazy
idiot (as everybody knows), but mostly because
[--------------------------------].

Also of note:  The sequence to add an alpha channel to an image in GIMP
is conveniently exactly the same as the options I use with `ls` (though
with `[LTRT]+[l]` instead of just `[l]`).  So that's convenient, I
think.