Ruined Everything
Hello.  I've just made a Tumblr post that also contains this info:
https://bluemanedhawk.tumblr.com/post/656201241705299968

I started this project a few months ago, and I had had the idea for it
long before then.  When I first began the project, I was deeply naive.
I thought that I would be able to create this game by having the
program draw to the screen directly, with support for X11 being a side
thing.  I though that I'd be able to deal with the differences
between every user's computer.  I thought that I'd somehow be able to
do 3D graphics in this project with zero knowledge on how to do so.
I thought I'd be able to create a heavy compression algorithm designed
specifically for this game to compress the assets.  Looking back, I
cannot believe how stupid I was.

Over time, I would eventually tone down my plans.  I decided to use
Simple DirectMedia Layer to draw to the screen within a window, and to
help deal with the differences between all the possible computers.  I
decided to keep the game 2D, with a vague possibility of 3D in the
future.  I ended up deciding to use BMPs and WAVs, since those options
were built in to SDL.  Nonetheless, I stayed adamant on a few things,
such as keeping the screen resolution as 4:3 480p, continuing with
exclusively using the keyboard, keeping with C with no game engine,
using simple digital audio to complement the low-resolution graphics,
having the game be plug-and-play, and not overcomplicating things with
multiple source files.

However, after that, I found myself unable to work on the project, and
I decided to put it on temporary hiatus.  I'm not going to go into
details as to why this is, partially because the details aren't really
that relevant to this, but mostly because it would require me to
disclose personal information.  What's important is that this caused
me to not have anywhere near enough energy to be able to work on this
project, but I nevertheless had plenty of time to think about it,
during which time I made some decisions that I deeply regret.

I have now resumed work on the project, and I have implemented the
regretable decisions I made:

- I decided to do mouse support.
- I ended up creating an `include/` directory and a Makefile.
- I decided to use recorded audio for the sound.
- I used some SDL extensions.
- I probably did some other things I'm forgetting, too.

However, there are still some things that I shall stand like a statue
on:

- I will continue to use C for this project.
- I will not use a game engine for this project.
- I will keep the screen resolution as 4:3 480p.
- I will keep the game fairly plug-and-play.
- I will probably also do some other things I'm forgetting.

I hope that I can still make _Ex Nihilo_ into the great game that I
believe it can be.

Thank you.