Began work on crash handler
I have setup a crash function and I have implemented it in several
places.  I have also defined a crash protocol in`docs/Problems/
CRASHES.md`.

My next plan is to set up a crash handler to deal with signals and to
flesh out the crash function a bit more.  Currently, however, I'm tired,
so that will need to wait.