Signal handling implemented
I've added the function `CrashHandlerSetup ( )` to `include/Crash.h`.
This is called at the beginning of the program and sets up the program
to deal with signals.  I've also added a .gitignore, made the makefile
look for a directory that actually exists when compressing the assets,
created a false macro in `include/Global.h`, and I probably also did
some other things I'm forgetting about.