Setup of SDL mostly complete
I still haven't figured out how to get FLAC to work with SDL_mixer.  It
appears that I may need to compile and ship a copy of it specifically
built with FLAC support, which is bloody great, because as we all know,
the point of dynamic libraries is that you need a different copy for
each program that uses it.

However, the game now has a window and a renderer, and will load some
particularly important assets when it starts.