Set up a skeleton and failed to get head to rotate
I cannot, for the life of me, figure out how to, determine the angle of
a vector from the center of the screen to the location of the cursor and
use it to rotate a texture located at the center of the screen to face
in the direction of the cursor.

Also, I set up a general structure for how I want the program to be laid
out and set up a bunch of dummy functions for it.

I haven't been able to do very much lately because real life has been
getting in the way during a period of time where I would logically have
had more free time to be able to work on this and more energy to be able
to work on this without procrastinating on it.  Sorry.